requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Sources
yfinance>=0.2.28
ccxt>=4.0.0
alpha_vantage>=2.3.1
requests>=2.31.0
aiohttp>=3.8.0
websockets>=11.0

# Technical Analysis
ta>=0.10.2
pandas-ta>=0.3.14b
tulipy>=0.4.0

# Backtesting
backtrader>=1.9.78.123
pyfolio>=0.9.2

# API & Dashboard
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
websockets>=11.0
plotly>=5.17.0
dash>=2.13.0
dash-bootstrap-components>=1.5.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0

# Async
asyncio-throttle>=1.0.2
aiofiles>=23.2.1

# Utils
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
loguru>=0.7.2
rich>=13.6.0
click>=8.1.7
apscheduler>=3.10.4
pytz>=2023.3

# Broker Integrations
alpaca-trade-api>=3.0.2

# Notifications
python-telegram-bot>=20.6
